further consolidate code into fewer files
