+++ src/index.ts
@@ -1,9 +1,3 @@
-/**
- * These are the public exports for this library available by importing/requiring:
- *
- * `import { run } from 'graphile-config-example'`
- */
-
 import {
   Middleware,
   GraphileConfig,
@@ -11,11 +5,28 @@ import {
   MiddlewareHandlers,
   resolvePreset,
 } from "graphile-config";
+import http from "node:http";
+import { z } from "zod";
 
-import { getDefaultPort, getServer } from "./get-server.js";
-import { ExampleMiddleware } from "./interfaces.js";
+import type {
+  ExampleMiddleware,
+  HandleRequestMiddlewareEvent,
+  MiddlewareContext,
+} from "./interfaces.js";
+import {
+  getCurrentTemperatureKelvin,
+  getDefaultAPIKey,
+} from "./open-weather.js";
+import {
+  respondWithBadRequest,
+  respondWithInternalServerError,
+  respondWithNotFound,
+  respondWithSuccess,
+  getNumberEnvironmentVariable,
+  constructRequestUrl,
+} from "./utils.js";
 
-export { BasicAuthenticationPreset } from "./config/presets/basic-authentication-preset.js";
+export { BasicAuthenticationPreset } from "./presets/basic-authentication-preset.js";
 
 declare global {
   namespace GraphileConfig {
@@ -81,4 +92,88 @@ function getExampleMiddleware(resolvedPreset: GraphileConfig.ResolvedPreset) {
     },
   );
   return middleware;
+}
+
+/**
+ * The use of Zod and Node's http are not related to Graphile Config. We use them
+ * simply to demonstrate how one might use Graphile Config in a project that runs
+ * an HTTP server.
+ *
+ * If you are creating a larger project, you may choose to use a more sophisticated
+ * HTTP server library like [Koa](https://koajs.com/) or
+ * [Express](https://expressjs.com/).
+ */
+
+const CURRENT_TEMPERATURE_SEARCH_PARAM_SCHEMA = z.object({
+  lat: z.coerce.number().finite(),
+  lon: z.coerce.number().finite(),
+});
+
+const getServer = (
+  resolvedPreset: GraphileConfig.ResolvedPreset,
+  middleware: Middleware<ExampleMiddleware>,
+): http.Server => {
+  const { openWeather: { apiKey = getDefaultAPIKey() } = {} } = resolvedPreset;
+  const context: MiddlewareContext = { resolvedPreset, apiKey };
+  const server = http.createServer((request, response) => {
+    const event: HandleRequestMiddlewareEvent = {
+      context,
+      request,
+      response,
+    };
+
+    // By using `runSync`, we require that all middleware functions are synchronous.
+    // If they need to do anything asynchronous, they should use next.callback().
+    //
+    // We could, instead, use `run()` if we had a use case that benefited from
+    // asynchronous middleware.
+    middleware.runSync("handleRequest", event, ({ request, response }) => {
+      const url = constructRequestUrl(request);
+      if (!url) {
+        respondWithInternalServerError(response);
+        return;
+      }
+
+      if (request.method === "GET" && url.pathname === "/current-temperature") {
+        handleCurrentTemperatureRequest(context, url, response);
+        return;
+      }
+
+      respondWithNotFound(response);
+    });
+  });
+  return server;
+};
+
+const handleCurrentTemperatureRequest = (
+  context: MiddlewareContext,
+  requestUrl: URL,
+  response: http.ServerResponse,
+): void => {
+  const parseResult = CURRENT_TEMPERATURE_SEARCH_PARAM_SCHEMA.safeParse(
+    Object.fromEntries(requestUrl.searchParams),
+  );
+
+  if (parseResult.success) {
+    getCurrentTemperatureKelvin(
+      context.apiKey,
+      Number(parseResult.data.lat),
+      Number(parseResult.data.lon),
+    )
+      .then((temperatureKelvin) => {
+        respondWithSuccess(response, { temperatureKelvin });
+      })
+      .catch((error: unknown) => {
+        console.error(error);
+        respondWithInternalServerError(response);
+      });
+  } else {
+    respondWithBadRequest(response, {
+      queryParams: parseResult.error,
+    });
+  }
+};
+
+function getDefaultPort() {
+  return getNumberEnvironmentVariable("PORT") ?? 4000;
 }
