+++ src/open-weather.ts
@@ -1,5 +1,20 @@
 import { z } from "zod";
 
+import { requireEnvironmentVariable } from "./utils.js";
+
+export const OPEN_WEATHER_API_KEY_ENVIRONMENT_VARIABLE_NAME =
+  "OPEN_WEATHER_API_KEY";
+const MISSING_OPEN_WEATHER_API_KEY_ERROR_MESSAGE =
+  "Missing OpenWeather API Key. Set it in your preset at the path openWeather.apiKey " +
+  `or in your environment variables as ${OPEN_WEATHER_API_KEY_ENVIRONMENT_VARIABLE_NAME}.`;
+
+export function getDefaultAPIKey() {
+  return requireEnvironmentVariable(
+    OPEN_WEATHER_API_KEY_ENVIRONMENT_VARIABLE_NAME,
+    MISSING_OPEN_WEATHER_API_KEY_ERROR_MESSAGE,
+  );
+}
+
 const OPEN_WEATHER_API_BASE_URL = "https://api.openweathermap.org";
 
 const CURRENT_WEATHER_API_PATH = "/data/2.5/weather";
