+++ src/plugins/http-basic-authentication-plugin.ts
