+++ src/presets/basic-authentication-preset.ts
@@ -1,5 +1,5 @@
-import { EnforceAuthenticationPlugin } from "../../plugins/authentication/enforce-authentication-plugin.js";
-import { BasicHttpAuthenticationPlugin } from "../../plugins/authentication/http-basic-authentication-plugin.js";
+import { EnforceAuthenticationPlugin } from "../plugins/enforce-authentication-plugin.js";
+import { BasicHttpAuthenticationPlugin } from "../plugins/http-basic-authentication-plugin.js";
 
 /**
  * Library authors may bundle together and export commonly used plugins as a
