+++ src/utils.ts
@@ -36,3 +36,44 @@ export const respondWithInternalServerError = (
     JSON.stringify({ errorMessage: "Something unexpected went wrong" }),
   );
 };
+
+export const constructRequestUrl = (
+  request: http.IncomingMessage,
+): URL | null => {
+  if (request.url && request.headers.host) {
+    return new URL(request.url, `http://${request.headers.host}`);
+  }
+
+  return null;
+};
+
+export const getNumberEnvironmentVariable = (
+  environmentVariableName: string,
+  rejectNaN = true,
+): number | undefined => {
+  const environmentVariable = process.env[environmentVariableName];
+  const parsed = environmentVariable
+    ? parseInt(environmentVariable, 10)
+    : undefined;
+
+  if (rejectNaN && parsed !== undefined && isNaN(parsed)) {
+    throw new Error(
+      `Environment variable is an invalid number: ${environmentVariableName}`,
+    );
+  }
+
+  return parsed;
+};
+
+export const requireEnvironmentVariable = (
+  variableName: string,
+  errorMessage?: string,
+): string => {
+  const value = process.env[variableName];
+  if (!value) {
+    throw new Error(
+      errorMessage ?? `Missing required environment variable: ${variableName}.`,
+    );
+  }
+  return value;
+};
